+++ Tests/Set-TeamviewerAccessToken.Tests.ps1
@@ -4,7 +4,7 @@ Import-Module $PSScriptRoot\..\Posh-Teamviewer\Posh-TeamViewer.psd1
 InModuleScope 'Posh-Teamviewer' {
 
     Describe 'Set-TeamviewerAccessToken' {
-        $AccessToken = 'Fake-AccessTokenText123456789'
+        $AccessToken = ConvertTo-SecureString -String 'Fake-AccessTokenText123456789' -AsPlainText -Force
         $MasterPassword = ConvertTo-SecureString -String 'FakePassword' -AsPlainText -Force
 
         Context 'AppData Teamviewer Folder Exists' {