+++ Tests/Update-TeamviewerDevice.Tests.ps1
@@ -5,7 +5,7 @@ Import-Module $PSScriptRoot\..\Posh-Teamviewer\Posh-TeamViewer.psd1
 InModuleScope 'Posh-Teamviewer' {
 
     Describe 'Update-TeamviewerDevice' {
-        $AccessToken = 'Fake-AccessTokenText123456789'
+        $AccessToken = ConvertTo-SecureString -String 'Fake-AccessTokenText123456789' -AsPlainText -Force
         $Password = 'FakePassword'
 
         Context 'Updates Device Description and Alias' {